requirements.txt
black
boto3==1.17.111
isort
moto==2.0.11
pandas==1.3.0
pyarrow==4.0.1
pytest==6.2.4
PyYAML==5.4.1
requests==2.26.0
